11/6/14 -1
Added daylogic file to process the days and times of the classes.
Updated get-times to work with that page. Next step is getting enough
test data set up to actually test the system.

+++ queries/get_times.php
@@ -11,24 +11,29 @@ white-space: pre;
 <?php
 //File to get the schedules
 include '../includes/connection.php';
-//include 'setup.php';
+//include 'daylogic.php';
+
 $array = array("COMP SCI   53 - Introduction To Programming",'COMP SCI  153 - Data Structures','COMP SCI  253 - Algorithms');
 $stunum = rand();
 $size = count($array);
+//echo $size.'<br>'; //TESTING OUTPUT
 $count = 0;//count should always equal the number of classes in a students schedule
 
 //Removes other schedules with the same student random number should they exist
 $q = "DELETE FROM `schedule` WHERE `stu_num_rand` = '$stunum'";
 mysql_query($q,$conm);
-$name = $array[0];
-echo 'Name: '.$name.'<br>';
+
+$name = $array[0];//gets the first class from the array passed to this page
+echo 'Name: '.$name.'<br>';//echos the name for testing purposes
+
+//gets the list of class numbers that correspond to this class name
 $q = "SELECT `number` FROM `class_num` WHERE `name` = '$name'";
 $res = mysql_query($q,$conm);
-$rcount = mysql_num_rows($res);//ERROR HERE NEED TO INVESTIGATE
-echo $q.'<br>';
-$rcount = 4;
-echo 'Rcount'.$rcount.'<br>';
+$rcount = mysql_num_rows($res); // counts the number of numbers 
+//echo $q.'<br>'; // echo for testing purposes
+//echo 'Rcount'.$rcount.'<br>';//echos for testing purposes
 
+//THIS SECTION CREATES THE FIRST SCHEDULE FROM THE LIST OF CLASS NUMBERS OF THE FIRST CLASS
 while($rcount > 0){
 	$rcount--;
 	$rnum = mysql_fetch_row($res);
@@ -38,7 +43,7 @@ while($rcount > 0){
 	echo $q.'<br>';
 	mysql_query($q,$conm);
 }
-$count++;
+$count++;//increases count to keep track of how many classes we have
 
 while($count < $size){
 	$name = $array[$count];
@@ -59,27 +64,60 @@ while($count < $size){
 		while($tempcount > 0){//get info from previous schedules
 			$tempcount--;
 			$car = array(0,0,0,0,0,0,0);
+			$days = array(0,0,0,0,0,0,0);
+			$stime = array(0,0,0,0,0,0,0);
+			$etime = array(0,0,0,0,0,0,0);
 			$restrow = mysql_fetch_row($rest);
 			$ttcount = 0;
 			while($ttcount < 7){
+				$tp3 = $ttcount+3;
 				$car[$ttcount] = $restrow[$ttcount+3];
+				if($restrow[$ttcount+3] != 0){
+					$qqr = "SELECT `start_time`,`end_time`,`days` FROM `classlist` WHERE `number` = '$restrow[$tp3]'";
+					echo $qqr.'<br>';
+					$resqr = mysql_query($qqr,$conm);
+					$rowqr = mysql_fetch_row($resqr);
+					$stime[$ttcount] = $rowqr[0];
+					$etime[$ttcount] = $rowqr[1];
+					$days[$ttcount] = $rowqr[2];
+				}
 				$ttcount++;
 			}
 			$car[$count] = $num;
-			//HERE WILL BE THE LOGIC TO CHECK if the schedule is valid, the q will go by a bool
-			//check time logic
-			//if start time 1 == start time 2 => same time
-			//if end time 1 == end time 2 => same time
-			//if start time 1 > start time 2 && end time 1 < end time 2 => same time -- ex geo lab, 1-3:50 this would apply to 2-2:50 classes
-			$q = "INSERT INTO `schedule`(stu_num_rand,class_num,class2_num,class3_num,class4_num,class5_num,class6_num,class7_num,num_classes)";
-			$q .= " VALUES('$stunum','$car[0]','$car[1]','$car[2]','$car[3]','$car[4]','$car[5]','$car[6]','$count')";
-			mysql_query($q,$conm);
-			echo $q.'<br>';
+			// HERE WILL BE THE LOGIC TO CHECK if the schedule is valid, the q will go by a bool
+			
+			// Gets the start time and end time of the class we are looking at
+			// need logic for days
+				$qn = "SELECT `start_time`,`end_time`,`days` FROM `classlist` WHERE `number` = '$num'";
+				echo $qn.'<br>';
+				$resqn = mysql_query($qn,$conm);
+				$nrow = mysql_fetch_row($resqn);
+				$nstime = $nrow[0];
+				$netime = $nrow[1];
+				$ndays = $nrow[2];
+				$tempc = $count;
+				echo 'Days and Times check<br>';
+				echo $nstime.' '.$netime.' '.$ndays.' '.$tempc.'<br>';
+				echo $stime[$ttcount].' '.$etime[$ttcount].' '.$days[$ttcount].'<br>';
+				 $ttcount = 0;
+				while($ttcount < 7){
+					$bool = check_days($ndays,$days[$ttcount],$nstime,$stime[$ttcount],$netime,$etime[$ttcount]);
+				}
+				
+			if($bool){
+				$q = "INSERT INTO `schedule`(stu_num_rand,class_num,class2_num,class3_num,class4_num,class5_num,class6_num,class7_num,num_classes)";
+				$q .= " VALUES('$stunum','$car[0]','$car[1]','$car[2]','$car[3]','$car[4]','$car[5]','$car[6]','$count')";
+				mysql_query($q,$conm);
+				echo $q.'<br>';
+			}
 				
 		}
 	}
-	$q = "DELETE FROM `schedule` WHERE `stu_num_rand` = '$stunum' AND `num_classes` < '$count'";
-	mysql_query($q,$conm);
+	if($count != $size){
+		$q = "DELETE FROM `schedule` WHERE `stu_num_rand` = '$stunum' AND `num_classes` < '$count'";
+		mysql_query($q,$conm);
+		echo 'DELETE STEP: '.$q.'<br>';
+	}
 
 	$count++;
 }
